rawdata-processing/requirements.txt
boto3==1.34
requests==2.32.3
python-dotenv==1.0
pandas==2.2
pydantic==2.7